move bip322 signing functions to testutil package
